+++ lib/src/client/options.dart
@@ -17,6 +17,11 @@ import 'dart:math';
 import 'transport/http2_credentials.dart';
 
 const defaultIdleTimeout = Duration(minutes: 5);
+
+/// It seems like Google's gRPC endpoints will forcefully close the
+/// connection after precisely 1 hour. So we *proactively* refresh our
+/// connection after 50 minutes. This will avoid one failed RPC call.
+const defaultConnectionTimeOut = Duration(minutes: 50);
 const defaultUserAgent = 'dart-grpc/2.0.0';
 
 typedef Duration BackoffStrategy(Duration lastBackoff);
@@ -39,16 +44,17 @@ Duration defaultBackoffStrategy(Duration lastBackoff) {
 class ChannelOptions {
   final ChannelCredentials credentials;
   final Duration idleTimeout;
+
+  /// The maximum time a single connection will be used for new requests.
+  final Duration connectionTimeout;
   final BackoffStrategy backoffStrategy;
   final String userAgent;
 
   const ChannelOptions({
-    ChannelCredentials credentials,
-    Duration idleTimeout,
-    String userAgent,
-    BackoffStrategy backoffStrategy,
-  })  : this.credentials = credentials ?? const ChannelCredentials.secure(),
-        this.idleTimeout = idleTimeout ?? defaultIdleTimeout,
-        this.userAgent = userAgent ?? defaultUserAgent,
-        this.backoffStrategy = backoffStrategy ?? defaultBackoffStrategy;
+    this.credentials = const ChannelCredentials.secure(),
+    this.idleTimeout = defaultIdleTimeout,
+    this.userAgent = defaultUserAgent,
+    this.backoffStrategy = defaultBackoffStrategy,
+    this.connectionTimeout = defaultConnectionTimeOut,
+  });
 }